backup/requirements.txt
requests
slack_sdk
kubernetes
psutil>=5.9.0
prometheus_client>=0.17.1
pandas>=2.0.0
schedule>=1.1.0
python-dotenv

